backend/pyproject.toml
[project]
name = "ndi-data-browser-v2-backend"
version = "2.0.0"
description = "NDI Data Browser v2 backend — cloud-first proxy + enricher"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "httpx[http2]>=0.27.0",
    "redis>=5.0.0",
    "cachetools>=5.4.0",
    "cryptography>=43.0.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.20.0",
    "python-multipart>=0.0.9",
    "Pillow>=10.4.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scipy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "respx>=0.22.0",
    "hypothesis>=6.110.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "fakeredis>=2.24.0",
    "httpx-ws>=0.7.0",
    "freezegun>=1.5.0",
]
observability = [
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "opentelemetry-instrumentation-httpx>=0.47b0",
    "sentry-sdk>=2.13.0",
]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "A", "C4",
    "SIM", "ARG", "PTH", "ERA", "PL", "RUF",
]
ignore = ["E501", "PLR0913", "PLR2004"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=. --cov-report=term-missing --cov-report=xml"
filterwarnings = ["error"]

[tool.coverage.run]
branch = true
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
fail_under = 85
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
